API / models: added `Nameable` concern
Provides a localizable name for a model.

Under the hood, `Message:Text` with a corresponding `lang` is used for
each locale the translation is needed for.

If the `lang` isn’t set (`nil`), the name is considered international
and applicable to any locale. It can be further overridden with a
locale-specific one though.

## Prerequisites

* `name_text_messages` association.

## Loading

It’s included automagically into any model having a `name_text_messages`
association.